[Bronze V] Title: 이상한 기호, Time: 56 ms, Memory: 5492 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -1,2 +1,3 @@
-int[] n = Array.ConvertAll(Console.ReadLine().Split(), int.Parse);
+long[] n = Array.ConvertAll(Console.ReadLine().Split(), long.Parse);
+
         Console.Write((n[0] + n[1]) * (n[0] - n[1]));
